requirements.txt
# Core dependencies
reflex==0.7.10
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic>=1.15.2

# Email and API services
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Google Cloud Storage
google-cloud-storage==2.10.0

# AI/ML services
google-generativeai==0.3.2

# Image processing for OCR
Pillow>=10.0.0

# Additional utilities
requests
python-multipart

# Reflex dependencies (auto-generated)
annotated-types
anyio
bidict


cachetools
certifi
charset-normalizer
click
ecdsa
fastapi
granian
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
Jinja2
jmespath
Mako
markdown-it-py
MarkupSafe
mdurl
oauthlib
packaging
platformdirs
proto-plus
psutil
pyasn1
pyasn1_modules
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
python-engineio
python-http-client
python-socketio
PyYAML
redis
reflex-hosting-cli
requests-oauthlib
rich
rsa
s3transfer
simple-websocket
six
sniffio
starlette
tabulate
tqdm
typing-inspection
typing_extensions
uritemplate
urllib3
uvicorn
watchfiles
Werkzeug
wrapt
wsproto>=1.2.0
